# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "URL Shortener",
        "description": "",
        "contact": {
            "email": "sachin.adkarr@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        }
    ],
...
    "paths": {
...
        "/short": {
            "post": {
                "tags": [
                    "ShortenUrl"
                ],
                "summary": "Generates a short URL",
                "operationId": "Shorten",
                "description": "",
                "requestBody": {
                  "content": {
                    "application/json": {
                      "schema": {
                        "$ref": "#/components/schemas/shortenUrlRequest"
                      }
                    }
                  }
                },

                "responses": {
                    "200": {
                        "description": "Success (OK)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/shortenUrlResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorCodes"
                                },
                                "examples": {
                                    "Please enter a url": {
                                      "value": {
                                         "code": "U01", "description": "Please enter a url!"
                                      }
                                    },
                                      "Enter a valid URL!": {
                                      "value": {
                                         "code": "U02", "description": "Enter a valid URL!"
                                      }
                                    },
                                    "Failed to generate short URL": {
                                      "value": {
                                         "code": "S03", "description": "Failed to generate short URL"
                                      }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorCodes"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "accessToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "accessToken": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "ErrorCodes": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "title": "Code",
                        "description": "Code of the error"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "Description of the error"
                    }
                }
            },
...
           "shortenUrlRequest":{
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string",
                        "title": "URL",
                        "description": "Url to be shortened"
                    }
                }
            },
            "shortenUrlResponse":{
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string",
                        "title": "URL",
                        "description": "shortened url"
                    }
                }
            }
        }
    }
}